main

+++ src/components/crud/Item.js
@@ -1,5 +1,5 @@
 import { createRoot } from 'react-dom/client';
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import { ReadItem } from './ReadItem';
 import { UpdateItem } from './UpdateItem';
 import { DeleteItem } from './DeleteItem';
@@ -12,23 +12,36 @@ export const Item = ({ classType, icons, item, urlApi, state }) => {
     const IconEdit = icons[classType].IconEdit;                                                               // Selección de icono update
     const IconDelete = icons[classType].IconDelete;                                                           // Selección de icono delete
 
-    console.log("open: ",open)
+    // console.log("open Item: ",open)
+
+    // console.log("state Item: ",state[0].value)
+
 
     const wideItems = ['paciente','doctor','consultorio','tratamiento','especialidad', 'genero', 'eps'];       // Wide columns
 
     ( open !== false ) ? document.getElementById('body').classList.add('noScroll') : document.getElementById('body').classList.remove('noScroll')   // No scroll when alerts are open
     
     let modal = '';
+    // console.log("item: ", item)
+
+    // useEffect(()=>{                                                               // Carga los valores del item seleccionado en el estado para su actualización
+    //     state.forEach((property,index) =>{ property.setState( Object.values(item[classType])[index] ) });
+    //   },[])
+
 
     if ( open !== false ) {
-        document.getElementById('root').insertAdjacentHTML('afterend',`<div id="modal"></div>`);
-        modal = createRoot( document.getElementById('modal') );
+        state.forEach((property,index) =>{ property.setState( Object.values(item[classType])[index] ) });
+        if( state[0].value !== '' ){                                // Avoid an empty render
+            document.getElementById('root').insertAdjacentHTML('afterend',`<div id="modal"></div>`);
+            modal = createRoot( document.getElementById('modal') );
 
-        switch (open) {
-            case 'read': modal.render( <ReadItem classType={classType} Icon={IconRead} item={item} setOpen={setOpen} />  ); break;
-            case 'update': modal.render( <UpdateItem classType={classType} Icon={IconEdit} item={item} urlApi={urlApi} setOpen={setOpen} icons={icons} state={state} /> ); break;
-            case 'delete': modal.render( <DeleteItem classType={classType} item={item} urlApi={urlApi} setOpen={setOpen} /> ); break;
-        }            
+            switch (open) {
+                case 'read': modal.render( <ReadItem classType={classType} Icon={IconRead} item={item} setOpen={setOpen} />  ); break;
+                case 'update': modal.render( <UpdateItem classType={classType} Icon={IconEdit} item={item} urlApi={urlApi} setOpen={setOpen} icons={icons} state={state} /> ); break;
+                case 'delete': modal.render( <DeleteItem classType={classType} item={item} urlApi={urlApi} setOpen={setOpen} /> ); break;
+            }
+        }
+
     }
 
     return (
